% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-desc.R
\docType{data}
\name{lag.sim}
\alias{lag.sim}
\title{lag.sim--simulated sequential data}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(lag.sim)
}
\description{
Data were simulated based on code found at:
https://stephens999.github.io/fiveMinuteStats/simulating_discrete_chains_1.html
}
\examples{
data(lag.sim)

}
\references{
https://stephens999.github.io/fiveMinuteStats/simulating_discrete_chains_1.html
#'
}
\keyword{datasets}
