% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lag_dat}
\alias{lag_dat}
\title{lag_dat}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{seqvar}{A vector containing sequential data--3 values: A, B, and C}
  \item{group}{group membership for the seqvar column--groups are 1 and 2}
  \item{linenum}{line numbers for the seqvar column}
}
}
\usage{
data(lag_dat)
}
\description{
Simulated Sequential Data
}
\details{
Data were simulated based on code at:
https://stephens999.github.io/fiveMinuteStats/simulating_discrete_chains_1.html
}
\examples{
data(lag_dat)

lagTables::trprobs(lag_dat, lagvar = "seqvar")
lagTables::trprobs(lag_dat, lagvar = "seqvar", laggroup="group")
lagTables::lagmodels(lag_dat, lagcol="seqvar", lagnum=2)
}
\references{
https://stephens999.github.io/fiveMinuteStats/simulating_discrete_chains_1.html
}
\keyword{datasets}
