% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagTables.R
\name{lagmodels}
\alias{lagmodels}
\title{Estimates log linear models for stationarity of sequential data
Requires a vector of sequential data
Can request estimates by group}
\usage{
lagmodels(
  d,
  lagcol,
  laggroup = "",
  title = "Log Linear Models for Stationarity",
  lagnum = 1
)
}
\arguments{
\item{d}{dataframe containing sequential vector}

\item{lagcol}{vector in dataframe containing sequential vector}

\item{laggroup}{vector in dataframe containing grouping variable.  Default is empty (no groups)}

\item{title}{caption for table}

\item{lagnum}{number of lags. Default is 1}
}
\value{
Table with log linear estimates for lag sequential data
}
\description{
Estimates log linear models for stationarity of sequential data
Requires a vector of sequential data
Can request estimates by group
}
