% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagTables.R
\name{shmodels}
\alias{shmodels}
\title{Tests stationarity/homogeneity across groups for sequential data
Requires a vector of sequential data
Requires a group variable
Tests first half of discussion against second half}
\usage{
shmodels(
  d,
  lagcol,
  laggroup = "",
  title = "Homogeniety Tests for Log Linear Models",
  lagnum = 1
)
}
\arguments{
\item{d}{dataframe containing sequential vector}

\item{lagcol}{vector in dataframe containing sequential vector}

\item{laggroup}{vector in dataframe containing grouping variable.  Default is empty (no groups)}

\item{title}{caption for table}

\item{lagnum}{number of lags. Default is 1}
}
\value{
Table with log linear estimates for lag sequential data
}
\description{
Tests stationarity/homogeneity across groups for sequential data
Requires a vector of sequential data
Requires a group variable
Tests first half of discussion against second half
}
