% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagTables.R
\name{trprobs}
\alias{trprobs}
\title{Create Transitional Probability Tables and Plots}
\usage{
trprobs(d, lagvar, laggroup = NULL,
  title = "Lag Sequential Descriptive Stats", lagnum = 1, plots = 0,
  dname = "")
}
\arguments{
\item{d}{dataframe that contains the sequential vector}

\item{lagvar}{the name of the vector that contains sequential data}

\item{laggroup}{vector that contains the grouping variables (default is NULL)}

\item{title}{specify title for tables and plots.}

\item{lagnum}{the number of lags to report.  Default is 1}

\item{plots}{request bar plots of standardized residuals. Default is 0. Set to 1 for plots.}

\item{dname}{provide a name for the data set for table caption}
}
\value{
Table with tables for the observed and expected frequencies, as well as the transitional probabilities and standardized residuals.  Bar plots if requested
}
\description{
This function works with a vector of sequential data and creates
html tables for the observed and expected frequencies, as well
as the transitional probabilities and standardized residuals
}
\details{
If requested, it will create bar plots of standardized residuals
for transitional probs (Lehmann-Willenbrock, 2013)

Finally, it will create tables and plots by group if requested

Default is one lag but higher-order lags can be requested
}
